+++ sdk/resourcemanager/compute/armcompute/models.go
@@ -368,7 +368,7 @@ type CapacityReservationGroupInstanceView struct {
 	// READ-ONLY; List of instance view of the capacity reservations under the capacity reservation group.
 	CapacityReservations []*CapacityReservationInstanceViewWithName
 
-	// READ-ONLY; List of the subscriptions that the capacity reservation group is shared with. Note: Minimum api-version: 2023-09-01.
+	// READ-ONLY; List of the subscriptions that the capacity reservation group is shared with. Note: Minimum api-version: 2024-03-01.
 	// Please refer to https://aka.ms/computereservationsharing for more details.
 	SharedSubscriptionIDs []*SubResourceReadOnly
 }
@@ -388,7 +388,7 @@ type CapacityReservationGroupProperties struct {
 	// Specifies the settings to enable sharing across subscriptions for the capacity reservation group resource. Pls. keep in
 	// mind the capacity reservation group resource generally can be shared across
 	// subscriptions belonging to a single azure AAD tenant or cross AAD tenant if there is a trust relationship established between
-	// the AAD tenants. Note: Minimum api-version: 2023-09-01. Please refer to
+	// the AAD tenants. Note: Minimum api-version: 2024-03-01. Please refer to
 	// https://aka.ms/computereservationsharing for more details.
 	SharingProfile *ResourceSharingProfile
 
@@ -2091,6 +2091,30 @@ type Extension struct {
 
 	// Extension Properties.
 	Properties *CloudServiceExtensionProperties
+}
+
+// FilterDefinition - Specifies the principal that can access the path, it's AND logic
+type FilterDefinition struct {
+	// Specifies the process path that can access the resource
+	CommandLine *string
+
+	// Specifies the process name that can access the resource
+	ProcessName *string
+
+	// Specifies whether only run as elevated process that can access the resource
+	RunAsElevated *string
+
+	// Specifies the user group id that can access the resource
+	UserGroupID *string
+
+	// Specifies the user group name that can access the resource
+	UserGroupName *string
+
+	// Specifies the user id that can access the resource
+	UserID *string
+
+	// Specifies the user name that can access the resource
+	UserName *string
 }
 
 // Gallery - Specifies information about the Shared Image Gallery that you want to create or update.
@@ -4070,19 +4094,62 @@ type ProximityPlacementGroupUpdate struct {
 	Tags map[string]*string
 }
 
-// ProxyAgentSettings - Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
+// ProxyAgentSettings - Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2024-03-01.
 type ProxyAgentSettings struct {
 	// Specifies whether ProxyAgent feature should be enabled on the virtual machine or virtual machine scale set.
 	Enabled *bool
+	Imds    *ProxyAgentSettingsImds
 
 	// Increase the value of this property allows user to reset the key used for securing communication channel between guest
 	// and host.
 	KeyIncarnationID *int32
+	Wireserver       *ProxyAgentSettingsWireserver
+}
 
-	// Specifies the mode that ProxyAgent will execute on if the feature is enabled. ProxyAgent will start to audit or monitor
-	// but not enforce access control over requests to host endpoints in Audit mode,
-	// while in Enforce mode it will enforce access control. The default value is Enforce mode.
+type ProxyAgentSettingsImds struct {
+	// Specify access control profile for IMDS endpoint
+	AccessControlProfile *ProxyAgentSettingsImdsAccessControlProfile
+
+	// Specifies the resource id on the form of
+	// /subscriptions/{SubscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Compute/galleries/{galleryName}/InVMAccessSettings/{settings}/versions/{version}
+	AccessControlProfileReferenceID *string
+
+	// Specify whether auto upgrade the accessControlProfile artifact if accessControlProfileReferenceId is specified
+	AutoUpgradeAccessControlProfile *bool
+}
+
+// ProxyAgentSettingsImdsAccessControlProfile - Specify access control profile for IMDS endpoint
+type ProxyAgentSettingsImdsAccessControlProfile struct {
+	// Specifies the mode that ProxyAgent will execute on. ProxyAgent will start to simulate and log access control over requests
+	// to IMDS endpoints in Audit mode, while in Enforce mode it will enforce access
+	// control. The default value is Disabled mode.
 	Mode *Mode
+
+	// Specify a dictionary of rules, key is URL path, value is an array of allowed list
+	Rules map[string][]*RuleDefinition
+}
+
+type ProxyAgentSettingsWireserver struct {
+	// Specify access control profile for wireserver endpoint
+	AccessControlProfile *ProxyAgentSettingsWireserverAccessControlProfile
+
+	// Specifies the resource id on the form of
+	// /subscriptions/{SubscriptionId}/resourceGroups/{ResourceGroupName}/providers/Microsoft.Compute/galleries/{galleryName}/InVMAccessSettings/{settings}/versions/{version}
+	AccessControlProfileReferenceID *string
+
+	// Specify whether auto upgrade the accessControlProfile artifact if accessControlProfileReferenceId is specified
+	AutoUpgradeAccessControlProfile *bool
+}
+
+// ProxyAgentSettingsWireserverAccessControlProfile - Specify access control profile for wireserver endpoint
+type ProxyAgentSettingsWireserverAccessControlProfile struct {
+	// Specifies the mode that ProxyAgent will execute on. ProxyAgent will start to simulate and log access control over requests
+	// to wireserver endpoints in Audit mode, while in Enforce mode it will enforce
+	// access control. Default is Audit mode.
+	Mode *Mode
+
+	// Specify a dictionary, key is URL path, value is an array of allowed list
+	Rules map[string][]*RuleDefinition
 }
 
 // ProxyOnlyResource - The ProxyOnly Resource model definition.
@@ -4427,7 +4494,7 @@ type ResourceSKUsResult struct {
 
 type ResourceSharingProfile struct {
 	// Specifies an array of subscription resource IDs that capacity reservation group is shared with. Note: Minimum api-version:
-	// 2023-09-01. Please refer to https://aka.ms/computereservationsharing for more
+	// 2024-03-01. Please refer to https://aka.ms/computereservationsharing for more
 	// details.
 	SubscriptionIDs []*SubResource
 }
@@ -4882,6 +4949,15 @@ type RollingUpgradeStatusInfoProperties struct {
 	RunningStatus *RollingUpgradeRunningStatus
 }
 
+// RuleDefinition - Specifies one rule
+type RuleDefinition struct {
+	// Specifies an array of principals that be allow to access the path
+	Filters []*FilterDefinition
+
+	// Dictionary of
+	QueryParameters map[string]*string
+}
+
 // RunCommandDocument - Describes the properties of a Run Command.
 type RunCommandDocument struct {
 	// REQUIRED; The VM run command description.
@@ -5146,7 +5222,7 @@ type SecurityProfile struct {
 	// Specifies the Managed Identity used by ADE to get access token for keyvault operations.
 	EncryptionIdentity *EncryptionIdentity
 
-	// Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
+	// Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2024-03-01.
 	ProxyAgentSettings *ProxyAgentSettings
 
 	// Specifies the SecurityType of the virtual machine. It has to be set to any specified value to enable UefiSettings. The
@@ -8277,8 +8353,8 @@ type VirtualMachineScaleSetVMProfile struct {
 	UserData *string
 
 	// READ-ONLY; Specifies the time in which this VM profile for the Virtual Machine Scale Set was created. Minimum API version
-	// for this property is 2023-09-01. This value will be added to VMSS Flex VM tags when
-	// creating/updating the VMSS VM Profile with minimum api-version 2023-09-01.
+	// for this property is 2024-03-01. This value will be added to VMSS Flex VM tags when
+	// creating/updating the VMSS VM Profile with minimum api-version 2024-03-01.
 	TimeCreated *time.Time
 }